+++ wiki/changelog.md
@@ -5,11 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## `0.2.0` [!badge variant="info" text="NEXT"]
+## `0.3.0` [!badge variant="info" text="NEXT"]
 
 Expected: unknown
 
-## `0.1.2` [!badge variant="info" text="LATEST"]
+## `0.2.0` [!badge variant="info" text="LATEST"]
+
+Released: 2021-10-02
+
+### Added
+
+- [x] Added `Pydantic` error support
+- [x] Added bytes support for string arguments
+
+### Changed
+
+- [x] Removed trailing `Exception` from error name in `error.exception`
+
+## `0.1.2`
 
 Released: 2021-09-24